+++ src/schemas/AuthorizationRequestPayloadVID1.schema.ts
@@ -251,6 +251,9 @@ export const AuthorizationRequestPayloadVID1SchemaObj = {
         "jwt_vc": {
           "$ref": "#/definitions/JwtObject"
         },
+        "jwt_vc_json": {
+          "$ref": "#/definitions/JwtObject"
+        },
         "jwt_vp": {
           "$ref": "#/definitions/JwtObject"
         },
@@ -615,6 +618,12 @@ export const AuthorizationRequestPayloadVID1SchemaObj = {
         "pattern": {
           "type": "string"
         },
+        "contains": {
+          "$ref": "#/definitions/FilterV1Base"
+        },
+        "items": {
+          "$ref": "#/definitions/FilterV1BaseItems"
+        },
         "type": {
           "type": "string"
         }
@@ -629,6 +638,111 @@ export const AuthorizationRequestPayloadVID1SchemaObj = {
         "number",
         "string"
       ]
+    },
+    "FilterV1Base": {
+      "type": "object",
+      "properties": {
+        "const": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "enum": {
+          "type": "array",
+          "items": {
+            "$ref": "#/definitions/OneOfNumberString"
+          }
+        },
+        "exclusiveMinimum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "exclusiveMaximum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "format": {
+          "type": "string"
+        },
+        "minLength": {
+          "type": "number"
+        },
+        "maxLength": {
+          "type": "number"
+        },
+        "minimum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "maximum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "not": {
+          "type": "object"
+        },
+        "pattern": {
+          "type": "string"
+        },
+        "contains": {
+          "$ref": "#/definitions/FilterV1Base"
+        },
+        "items": {
+          "$ref": "#/definitions/FilterV1BaseItems"
+        },
+        "type": {
+          "type": "string"
+        }
+      },
+      "additionalProperties": false
+    },
+    "FilterV1BaseItems": {
+      "type": "object",
+      "properties": {
+        "const": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "enum": {
+          "type": "array",
+          "items": {
+            "$ref": "#/definitions/OneOfNumberString"
+          }
+        },
+        "exclusiveMinimum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "exclusiveMaximum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "format": {
+          "type": "string"
+        },
+        "minLength": {
+          "type": "number"
+        },
+        "maxLength": {
+          "type": "number"
+        },
+        "minimum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "maximum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "not": {
+          "type": "object"
+        },
+        "pattern": {
+          "type": "string"
+        },
+        "contains": {
+          "$ref": "#/definitions/FilterV1Base"
+        },
+        "items": {
+          "$ref": "#/definitions/FilterV1BaseItems"
+        },
+        "type": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "type"
+      ],
+      "additionalProperties": false
     },
     "HolderSubject": {
       "type": "object",
@@ -832,6 +946,141 @@ export const AuthorizationRequestPayloadVID1SchemaObj = {
         },
         "type": {
           "type": "string"
+        },
+        "contains": {
+          "$ref": "#/definitions/FilterV2Base"
+        },
+        "items": {
+          "$ref": "#/definitions/FilterV2BaseItems"
+        }
+      },
+      "required": [
+        "type"
+      ],
+      "additionalProperties": false
+    },
+    "FilterV2Base": {
+      "type": "object",
+      "properties": {
+        "const": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "enum": {
+          "type": "array",
+          "items": {
+            "$ref": "#/definitions/OneOfNumberString"
+          }
+        },
+        "exclusiveMinimum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "exclusiveMaximum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "format": {
+          "type": "string"
+        },
+        "formatMaximum": {
+          "type": "string"
+        },
+        "formatMinimum": {
+          "type": "string"
+        },
+        "formatExclusiveMaximum": {
+          "type": "string"
+        },
+        "formatExclusiveMinimum": {
+          "type": "string"
+        },
+        "minLength": {
+          "type": "number"
+        },
+        "maxLength": {
+          "type": "number"
+        },
+        "minimum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "maximum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "not": {
+          "type": "object"
+        },
+        "pattern": {
+          "type": "string"
+        },
+        "type": {
+          "type": "string"
+        },
+        "contains": {
+          "$ref": "#/definitions/FilterV2Base"
+        },
+        "items": {
+          "$ref": "#/definitions/FilterV2BaseItems"
+        }
+      },
+      "additionalProperties": false
+    },
+    "FilterV2BaseItems": {
+      "type": "object",
+      "properties": {
+        "const": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "enum": {
+          "type": "array",
+          "items": {
+            "$ref": "#/definitions/OneOfNumberString"
+          }
+        },
+        "exclusiveMinimum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "exclusiveMaximum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "format": {
+          "type": "string"
+        },
+        "formatMaximum": {
+          "type": "string"
+        },
+        "formatMinimum": {
+          "type": "string"
+        },
+        "formatExclusiveMaximum": {
+          "type": "string"
+        },
+        "formatExclusiveMinimum": {
+          "type": "string"
+        },
+        "minLength": {
+          "type": "number"
+        },
+        "maxLength": {
+          "type": "number"
+        },
+        "minimum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "maximum": {
+          "$ref": "#/definitions/OneOfNumberString"
+        },
+        "not": {
+          "type": "object"
+        },
+        "pattern": {
+          "type": "string"
+        },
+        "type": {
+          "type": "string"
+        },
+        "contains": {
+          "$ref": "#/definitions/FilterV2Base"
+        },
+        "items": {
+          "$ref": "#/definitions/FilterV2BaseItems"
         }
       },
       "required": [